refactor(mail-account): move BaseAccount to :feature:mail:account:api kmp module
