+++ app-common/build.gradle.kts
@@ -17,12 +17,11 @@ dependencies {
     implementation(projects.legacy.core)
     implementation(projects.legacy.account)
 
-    implementation(projects.core.account)
-
     implementation(projects.core.featureflags)
     implementation(projects.core.ui.legacy.theme2.common)
 
     implementation(projects.feature.account.setup)
+    implementation(projects.feature.mail.account.api)
     implementation(projects.feature.migration.provider)
     implementation(projects.feature.widget.messageList)
