+++ backend/api/build.gradle.kts
@@ -4,8 +4,8 @@ plugins {
 }
 
 dependencies {
-    implementation(projects.core.account)
     implementation(projects.core.outcome)
     implementation(projects.core.mail.folder.api)
+    implementation(projects.feature.mail.account.api)
     api(projects.mail.common)
 }
