+++ backend/api/src/main/kotlin/net/thunderbird/backend/api/BackendFactory.kt
@@ -1,7 +1,7 @@
 package net.thunderbird.backend.api
 
 import com.fsck.k9.backend.api.Backend
-import net.thunderbird.core.account.BaseAccount
+import net.thunderbird.feature.mail.account.api.BaseAccount
 
 interface BackendFactory<TAccount : BaseAccount> {
     fun createBackend(account: TAccount): Backend
