+++ backend/api/src/main/kotlin/net/thunderbird/backend/api/BackendStorageFactory.kt
@@ -1,7 +1,7 @@
 package net.thunderbird.backend.api
 
 import com.fsck.k9.backend.api.BackendStorage
-import net.thunderbird.core.account.BaseAccount
+import net.thunderbird.feature.mail.account.api.BaseAccount
 
 interface BackendStorageFactory<in TAccount : BaseAccount> {
     fun createBackendStorage(account: TAccount): BackendStorage
