+++ backend/imap/build.gradle.kts
@@ -6,7 +6,9 @@ plugins {
 dependencies {
     api(projects.backend.api)
     api(projects.core.outcome)
-    api(projects.core.account)
+
+    api(projects.feature.mail.account.api)
+
     api(projects.mail.protocols.imap)
     api(projects.mail.protocols.smtp)
