--- core/account/build.gradle.kts
@@ -1,4 +1,0 @@
-plugins {
-    id(ThunderbirdPlugins.Library.jvm)
-    alias(libs.plugins.android.lint)
-}
