+++ feature/mail/account/api/src/commonMain/kotlin/net/thunderbird/feature/mail/account/api/BaseAccount.kt
@@ -1,4 +1,4 @@
-package net.thunderbird.core.account
+package net.thunderbird.feature.mail.account.api
 
 interface BaseAccount {
     val uuid: String
