+++ feature/navigation/drawer/dropdown/build.gradle.kts
@@ -16,9 +16,9 @@ dependencies {
     implementation(projects.core.ui.compose.designsystem)
 
     implementation(projects.feature.account.avatar)
+    implementation(projects.feature.mail.account.api)
 
     implementation(projects.feature.search)
-    implementation(projects.core.account)
     implementation(projects.legacy.account)
     implementation(projects.legacy.mailstore)
     implementation(projects.legacy.message)
