+++ legacy/account/src/main/java/app/k9mail/legacy/account/LegacyAccount.kt
@@ -5,8 +5,8 @@ import com.fsck.k9.mail.Address
 import com.fsck.k9.mail.ServerSettings
 import java.util.Calendar
 import java.util.Date
-import net.thunderbird.core.account.BaseAccount
 import net.thunderbird.core.mail.folder.api.SpecialFolderSelection
+import net.thunderbird.feature.mail.account.api.BaseAccount
 import net.thunderbird.feature.notification.NotificationSettings
 
 // This needs to be in sync with K9.DEFAULT_VISIBLE_LIMIT
