+++ legacy/account/src/main/java/app/k9mail/legacy/account/LegacyAccountWrapper.kt
@@ -2,8 +2,8 @@ package app.k9mail.legacy.account
 
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.NO_OPENPGP_KEY
 import com.fsck.k9.mail.ServerSettings
-import net.thunderbird.core.account.BaseAccount
 import net.thunderbird.core.mail.folder.api.SpecialFolderSelection
+import net.thunderbird.feature.mail.account.api.BaseAccount
 import net.thunderbird.feature.notification.NotificationSettings
 
 /**
