+++ legacy/core/build.gradle.kts
@@ -17,7 +17,7 @@ dependencies {
     api(projects.feature.account.storage.legacy)
 
     api(projects.feature.search)
-    api(projects.core.account)
+    api(projects.feature.mail.account.api)
     api(projects.legacy.account)
     api(projects.legacy.di)
     api(projects.legacy.mailstore)
