+++ legacy/mailstore/build.gradle.kts
@@ -7,13 +7,14 @@ android {
 }
 
 dependencies {
-    implementation(projects.feature.search)
-    implementation(projects.core.account)
     implementation(projects.legacy.account)
     implementation(projects.legacy.di)
     implementation(projects.legacy.message)
 
-    implementation(projects.mail.common)
     implementation(projects.core.mail.folder.api)
+    implementation(projects.feature.mail.account.api)
     implementation(projects.feature.folder.api)
+    implementation(projects.feature.search)
+
+    implementation(projects.mail.common)
 }
