+++ settings.gradle.kts
@@ -130,6 +130,10 @@ include(
 )
 
 include(
+    ":feature:mail:account:api",
+)
+
+include(
     ":core:common",
     ":core:featureflags",
     ":core:outcome",
@@ -223,7 +227,6 @@ include(":core:android:logging")
 include(":core:preferences")
 include(":core:mail:mailserver")
 include(":feature:search")
-include(":core:account")
 include(":feature:notification")
 include(":core:ui:theme:manager")
 include(":core:contact")
